pyproject.toml
[project]
name = "mitra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=6.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
